requirements.txt
fastapi
uvicorn
Pillow
tensorflow==2.18.0
python-multipart
python-dotenv
google-generativeai